feat(theme): add image shortcode and use on about page
