+++ themes/mizzi/tests/build.test.ts
@@ -1,180 +1,5 @@
-import { test as base, expect } from '@playwright/test';
-import * as fs from 'fs';
-import * as path from 'path';
-import * as os from 'os';
-import { spawn } from 'child_process';
-import { fileURLToPath } from 'url';
-import getPort from 'get-port';
-
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
-const THEME_SOURCE_DIR = path.join(__dirname, '..', '..');
-const DEBUG_MODE = process.env.DEBUG_TESTS === 'true';
-
-// Server fixture interface
-interface ServerFixture {
-  start: (config?: TestSiteConfig) => Promise<{ baseURL: string }>;
-}
-
-// Test fixtures
-interface TestFixtures {
-  server: ServerFixture;
-}
-
-interface TestSiteConfig {
-  params?: string;
-  indexContent?: string;
-}
-
-interface TestSite {
-  tmpDir: string;
-  port: number;
-  cleanup: () => void;
-}
-
-interface ServerControl {
-  port: number;
-  baseURL: string;
-  stop: () => void;
-}
-
-async function createTestSite(config: TestSiteConfig = {}): Promise<TestSite> {
-  const tmpDir = fs.mkdtempSync(path.join(os.tmpdir(), 'hugo-theme-test-'));
-
-  fs.mkdirSync(path.join(tmpDir, 'config'));
-  fs.mkdirSync(path.join(tmpDir, 'config', '_default'));
-
-  const port = await getPort();
-  const defaultConfig = {
-    baseURL: `http://localhost:${port}`,
-    languageCode: 'en-us',
-    title: 'Theme Test Site',
-  };
-
-  const hugoConfig = Object.entries(defaultConfig)
-    .map(([key, value]) => `${key} = "${value}"`)
-    .join('\n');
-
-  fs.writeFileSync(path.join(tmpDir, 'config', '_default', 'hugo.toml'), hugoConfig);
-
-  // Write params.toml if provided
-  if (config.params) {
-    fs.writeFileSync(path.join(tmpDir, 'config', '_default', 'params.toml'), config.params);
-  }
-
-  // Create content directory and files if provided
-  if (config.indexContent) {
-    fs.mkdirSync(path.join(tmpDir, 'content'));
-    fs.writeFileSync(path.join(tmpDir, 'content', '_index.md'), config.indexContent);
-  }
-
-  return {
-    tmpDir,
-    port,
-    cleanup: () => {
-      if (!DEBUG_MODE) {
-        fs.rmSync(tmpDir, { recursive: true, force: true });
-      }
-      else {
-        console.log(`Debug mode: Test site preserved at ${tmpDir}`);
-      }
-    },
-  };
-};
-
-async function startHugoServer(testSite: TestSite): Promise<ServerControl> {
-  return new Promise((resolve, reject) => {
-    const server = spawn('hugo', [
-      'server',
-      '--disableFastRender',
-      '--ignoreCache',
-      '--port', testSite.port.toString(),
-      '--bind', '127.0.0.1',
-      '--theme', 'mizzi',
-      '--themesDir', THEME_SOURCE_DIR,
-    ], {
-      cwd: testSite.tmpDir,
-      stdio: 'pipe',
-    });
-
-    let serverStarted = false;
-
-    server.stdout?.on('data', (data: Buffer) => {
-      const output = data.toString();
-      console.log(`Hugo server output: ${output}`);
-
-      // Look for indication that server has started
-      if (output.includes('Web Server is available at') || output.includes('Press Ctrl+C to stop')) {
-        if (!serverStarted) {
-          serverStarted = true;
-          resolve({
-            port: testSite.port,
-            baseURL: `http://localhost:${testSite.port}`,
-            stop: () => {
-              server.kill();
-            },
-          });
-        }
-      }
-    });
-
-    server.stderr?.on('data', (data: Buffer) => {
-      console.error(`Hugo server error: ${data}`);
-    });
-
-    server.on('error', (error: Error) => {
-      reject(error);
-    });
-
-    server.on('exit', (code: number | null) => {
-      if (code !== 0 && !serverStarted) {
-        reject(new Error(`Hugo server exited with code ${code}`));
-      }
-    });
-  });
-}
-
-// Global cleanup tracking
-const activeServers: ServerControl[] = [];
-const activeTestSites: TestSite[] = [];
-
-// Extended test with server fixture
-const test = base.extend<TestFixtures>({
-  // eslint-disable-next-line no-empty-pattern
-  server: async ({ }, use) => {
-    const serverFixture: ServerFixture = {
-      start: async (config: TestSiteConfig = {}) => {
-        const testSite = await createTestSite(config);
-        const serverControl = await startHugoServer(testSite);
-
-        // Track for cleanup
-        activeTestSites.push(testSite);
-        activeServers.push(serverControl);
-
-        return { baseURL: serverControl.baseURL };
-      },
-    };
-
-    await use(serverFixture);
-  },
-});
-
-// Global cleanup
-test.afterEach(async () => {
-  // Stop all servers
-  for (const server of activeServers) {
-    server.stop();
-  }
-
-  // Cleanup all test sites
-  for (const site of activeTestSites) {
-    site.cleanup();
-  }
-
-  // Clear tracking arrays
-  activeServers.length = 0;
-  activeTestSites.length = 0;
-});
+import { expect } from '@playwright/test';
+import { test } from './fixtures.js';
 
 test.describe('Hugo Theme Tests', () => {
   test('should serve the theme homepage correctly', async ({ page, server }) => {
